backend/requirements.txt
Flask
flask-cors
python-dotenv
Pillow
google-genai
requests
gunicorn